pyproject.toml
[tool.poetry]
name = "private-assistant-time-skill"
version = "0.1.6"
description = "Time Management Skill for the Private Assistant: Handles timers, time telling, and other time-related functionalities."
authors = ["stkr22 <stkr22@github.com>"]
readme = "README.md"
license = "GNU GPLv3"
include = ["src"]

[tool.poetry.dependencies]
spacy = "^3.7.4"
typer = {version = "0.9.0", extras = ["all"]}
pydantic = "^2.7.0"
python = ">=3.11,<3.12"
pyyaml = "^6.0.1"
jinja2 = "^3.1.3"
en-core-web-md = {url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.7.1/en_core_web_md-3.7.1-py3-none-any.whl"}
sqlmodel = "^0.0.16"
psycopg2-binary = "^2.9.9"
private-assistant-commons = "^0.1.0"
word2number = "^1.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.3.7"
mypy = "^1.9.0"
pytest = "^7.4.4"
types-pyyaml = "^6.0.12.20240311"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
private-assistant-time-skill = "private_assistant_time_skill.main:app"

[tool.ruff]
target-version = "py311"
line-length = 88
indent-width = 4

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I", "UP"]
ignore = ["E203"]
